requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
google-genai>=1.0.0
aiosqlite>=0.20.0
pydantic>=2.0.0
